# src/main/resources/swagger/api_cuentas_banca_v1.yml
openapi: 3.0.3
info:
  title: Swagger Bank Management - OpenAPI 3.0
  description: |-
    This is a bank customer account management server, where the customer can use and consult their banking products.

  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://managementBank.swagger.io/api/v3
...
paths:
...
  /product/createAccount:
    put:
      tags:
        - product
      summary: Create account bank
      description: Create account bank
      operationId: registerAccount
      requestBody:
        description: Create account bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/AccountRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AccountRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoAccountResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductoAccountResponse'
        '405':
          description: Invalid input
  /product/associatedDebitCard:
    put:
      tags:
        - product
      summary: Associate debit card with account number
      description: Associate debit card
      operationId: associateDebitCard
      requestBody:
        description: Associate debit card
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociateRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/AssociateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AssociateRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssociateResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/AssociateResponse'
        '405':
          description: Invalid input
  /product/balance/account/{number}:
    get:
      tags:
        - product
      summary: Gets balance of client account number
      description: ets balance of client account number
      operationId: getBalanceProduct
      parameters:
        - name: number
          in: path
          description: Account number of product of client
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceAccountResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/BalanceAccountResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: product not found
  /product/balance/debit/{numberCard}:
    get:
      tags:
        - product
      summary: Gets balance of debit card
      description: Gets balance of debit card
      operationId: getBalanceDebit
      parameters:
        - name: numberCard
          in: path
          description: card number debit number of product of client
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceAccountResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/BalanceAccountResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: product not found
  /product/balance/credit/{numberCard}:
    get:
      tags:
        - product
      summary: Gets balance of credit card
      description: Gets balance of credit card
      operationId: getBalanceCredit
      parameters:
        - name: numberCard
          in: path
          description: card number credit of product of client
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceAccountResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/BalanceAccountResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: product not found
  /product/balance/{numberDocument}:
    get:
      tags:
        - product
      summary: Gets balance all product of client
      description: Gets balance all product of client
      operationId: getAllProduct
      parameters:
        - name: numberDocument
          in: path
          description: number document of client
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductBalanceResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProductBalanceResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: product not found
  /product/movements/{number}:
    get:
      tags:
        - product
      summary: Gets the movements of client account or credit card
      description: Gets the movements of client account or credit card
      operationId: getMovements
      parameters:
        - name: number
          in: path
          description: Account number or Credit number of product of client
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceMovementsResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/BalanceMovementsResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: client not found
...
components:
  schemas:
...
    DataPersonal:
      type: object
      properties:
        documentType:
          type: string
          example: "1"
        documentNumber:
          type: string
          example: "45859966"
        name:
          type: string
          example: "Jorge"
        lastFather:
          type: string
          example: "Velazco"
        lastMother:
          type: string
          example: "Peralta"
        birthdate:
          type: string
          example: "1985-10-12"
      xml:
        name: dataPersonal
      required:
        - documentType
        - documentNumber
        - name
        - lastFather
        - lastMother
        - birthdate
    DataCompany:
      type: object
      properties:
        ruc:
          type: string
          example: "10458899999"
        name:
          type: string
          example: "Cementos Lima SAC"
        dateCreationCompany:
          type: string
          example: "2019-10-22"
      xml:
        name: dataCompany
      required:
        - ruc
        - name
        - dateCreationCompany
...
    Movements:
      type: object
      properties:
        numberOperation:
          type: string
          example: "10225"
        dateOperation:
          type: string
          example: "2023-01-10 10:05:45"
        operationType:
          type: string
          description: "Tipo de Operacion"
          example: "Retiro"
          enum:
            - RETIRO
            - DEPOSITO
            - CONSUMO
        comission:
          type: number
          example: 2.05
        grossAmount:
          type: number
          example: 40.00
        amount:
          type: number
          example: 2.05
      xml:
        name: movements
      required:
        - operationType
        - amount
...
    Sucursal:
      type: object
      properties:
        id:
          type: string
          example: "1544"
        name:
          type: string
          example: "Sucurcal Las Camelias"
        address:
          type: string
          example: "Calle Las Camelias 651"
        nameOperator:
          type: string
          example: "Patricio Franco Mena"
        nameManager:
          type: string
          example: "Susana Rios Torres"
      xml:
        name: sucursal
      required:
        - id
        - name
        - address
        - nameOperator
        - nameManager
...
    AccountRequest:
      type: object
      properties:
        numberDocument:
          type: string
          example: "45692341"
        clientType:
          type: string
          example: "Personal"
        profileType:
          type: string
          example: "VIP"
        accountType:
          type: string
          example: "Ahorro"
        sucursal:
          $ref: '#/components/schemas/Sucursal'
        amount:
          type: number
          example: 120.00
      xml:
        name: accountRequest
      required:
        - numberDocument
        - accountType
        - profileType
        - sucursal
        - amount
    ProductoAccountResponse:
      type: object
      properties:
        clientType:
          type: string
          example: "Personal"
        profileType:
          type: string
          example: "VIP"
        dataCompany:
          $ref: '#/components/schemas/DataCompany'
        dataPersonal:
          $ref: '#/components/schemas/DataPersonal'
        product:
          $ref: '#/components/schemas/AccountResponse'
      xml:
        name: productoAccountResponse
    AccountResponse:
      type: object
      properties:
        type:
          type: string
          example: "Ahorro"
        number:
          type: string
          example: "1992001451121"
        movements:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Movements'
        sucursal:
          $ref: '#/components/schemas/Sucursal'
        balance:
          type: number
          example: 3000.00
      xml:
        name: accountResponse
    BalanceAccountResponse:
      type: object
      properties:
        typeAccount:
          type: string
          example: "Ahorro"
        numberCard:
          type: string
          example: "4552758596322526"
        numberAccount:
          type: string
          example: "1992001451121"
        balance:
          type: number
          example: 3000.00
      xml:
        name: balanceAccountResponse
    BalanceCreditResponse:
      type: object
      properties:
        type:
          type: string
          example: "Personal"
        card:
          type: string
          example: "4552758596322526"
        number:
          type: string
          example: "1992001451121"
        limitCredit:
          type: number
          example: 8000.00
        balance:
          type: number
          example: 3000.00
      xml:
        name: balanceCreditResponse
    BalanceDebitResponse:
      type: object
      properties:
        card:
          type: string
          example: "4552758596322526"
        balance:
          type: number
          example: 3000.00
      xml:
        name: balanceDebitResponse
    ProductBalanceResponse:
      type: object
      properties:
        account:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/BalanceAccountResponse'
        credit:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/BalanceCreditResponse'
        debit:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/BalanceDebitResponse'
        totalBalance:
          type: number
          example: 3000.00
      xml:
        name: productBalanceResponse
    BalanceMovementsResponse:
      type: object
      properties:
        type:
          type: string
          example: "Ahorro"
        card:
          type: string
          example: "4552758596322526"
        number:
          type: string
          example: "1992001451121"
        movements:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Movements'
        balance:
          type: number
          example: 3000.00
      xml:
        name: balanceMovementsResponse
    AssociateRequest:
      type: object
      properties:
        documentNumber:
          type: string
          example: "45879965"
        debitCard:
          type: string
          example: "455785264586"
        accountAssociate:
          $ref: '#/components/schemas/AssociateAccountRequest'
      xml:
        name: associateRequest
      required:
        - documentNumber
        - debitCard
        - accountAssociate
    AssociateAccountRequest:
      type: object
      properties:
        accountNumber:
          type: string
          example: "1991456912568"
        acountType:
          type: string
          example: "Ahorro"
      xml:
        name: associateAccountRequest
      required:
        - accountNumber
        - acountType
    AssociateResponse:
      type: object
      properties:
        clientType:
          type: string
          example: "Personal"
        profileType:
          type: string
          example: "VIP"
        dataCompany:
          $ref: '#/components/schemas/DataCompany'
        dataPersonal:
          $ref: '#/components/schemas/DataPersonal'
        debit:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/DebitResponse'
      xml:
        name: productoAccountResponse
    DebitResponse:
      type: object
      properties:
        card:
          type: string
          example: "455785264586"
        assciateAccount:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/AccountDebitResponse'
        balance:
          type: number
          example: 3000.00
      xml:
        name: debitResponse
    AccountDebitResponse:
      type: object
      properties:
        accountNumber:
          type: string
          example: "1991456912568"
        dateCreation:
          type: string
          example: "2023-02-15 18:35:25"
      xml:
        name: accountDebitResponse